Add files via upload
This project was focused on using random forest classifier python to develop a loan prediction model. This model is to help financial institutions determine which  clients or potential clients can be recommended for loans based on set parameters. These parameters are to ensure that the clients who are offered loans do not pose threats as ;loan defaulters'. These parameters were modeled after based on the needs of a pseudo financial institution. However this model can be adjusted to include parameters that meet the needs of an institution. This was the first version of this model and can be improved to address any limitations of this first model.